+++ packages/bot-engine/whatsapp/WhatsappComponentFlow/sendWhatsappSocketMessage.ts
@@ -2,7 +2,7 @@ import { TRPCError } from '@trpc/server'
 import { env } from '@typebot.io/env'
 import { Socket, io } from 'socket.io-client'
 import { WhatsappOperationTypes } from './WhatsappOperationType'
-import { WhatsappSocketSendingMessage } from './convertMessageToWhatsappCompoent'
+import { WhatsappSocketSendingMessage } from './convertMessageToWhatsappComponent'
 import { getSession } from '../../queries/getSession'
 import prisma from '@typebot.io/lib/prisma'
 import axios from 'axios'
